refactored credential into verify master password form
